backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp==3.8.6
pydantic==2.4.2
httpx==0.25.1
pytest==7.4.3
selenium==4.15.2
webdriver-manager==4.0.1
fake-useragent==1.3.0
